+ Make basic like request to backend  + Error handled on frontend  + Factored out api controllers
Relates #99
